requirements.txt
django>=1.10.2
# mysql-python  # command: easy_install mysql-python
django-mysql
django-tastypie
docutils
mongoengine==0.10.6
# django-tastypie-mongoengine: pip install git+https://github.com/gagan144/django-tastypie-mongoengine  # enable text search: in /etc/mongodb.cnf, add: setParameter = textSearchEnabled=true
pymongo>=3.3.0
# NOT USING: django-mongoengine>=0.2.1 : pip install git+https://github.com/MongoEngine/django-mongoengine

blinker
ujson
django-tinymce
django-easy-select2
django-extensions
django-fsm
django-fsm-log
pyjwt
django-recaptcha
jsonobject
boto
django-storages
shortuuid
django-fcm
django-colorful
django-tree-tag
# Pillow



# ----- Documentation -----
# sphinx
# sphinx_rtd_theme